Split out auto-detection behaviour
